feat(types): changed structure of types
