+++ frontend/src/types/board/BoardData.tsx
@@ -1,0 +1,10 @@
+import { difficulty } from "@/types/search/difficulty";
+import { BoardType } from "@/types/board/BoardType";
+
+type BoardData = {
+  name: string;
+  difficulty: difficulty;
+  board: BoardType;
+};
+
+export type { BoardData };
