+++ frontend/src/types/board/BoardType.tsx
@@ -1,0 +1,3 @@
+type BoardType = string[][];
+
+export type { BoardType };
