+++ frontend/src/types/search/updatedAt.tsx
@@ -1,0 +1,11 @@
+type updatedAtType = "24hours" | "7days" | "1month" | "3months";
+
+enum updatedAt {
+  "24hours" = "24hours",
+  "7days" = "7days",
+  "1month" = "1month",
+  "3months" = "3months",
+}
+
+export type { updatedAtType };
+export { updatedAt };
